% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivCIF.R
\name{BivCIF}
\alias{BivCIF}
\title{Output estimate for the bivariate cumulative incidence function}
\usage{
BivCIF(BivHaz, BivSurv)
}
\arguments{
\item{BivHaz}{A matrix with estimates of bivariate cause specific hazard function. Times for variable 1 are along the rows, times for variable 2 along columns.}

\item{BivSurv}{A matrix with estimates of bivariate survival times. Times for variable 1 are along the rows, times for variable 2 along columns.}
}
\value{
A list with the following elements:
\itemize{
\item{BivIF: Jumps in the bivariate cause specific cumulative incidence function estimate}
\item{BivCIF: The bivariate cause specific cumulative incidence function estimate}
}
}
\description{
Takes bivariate survival and bivariate hazard function and outputs estimate for bivariate cumulative incidence function
}
