% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivDabrowska.R
\name{BivDabrowska}
\alias{BivDabrowska}
\title{Output the Dabrowska bivariate survival function estimate}
\usage{
BivDabrowska(data)
}
\arguments{
\item{data}{An n by 4 matrix where the first column is the observed event time of the first variable, the second column is the observed event time of the second variable, the third column is the censoring indicator for the first variable (0 indicates censored) and the fourth column is the censoring indicator for the second variable (0 indicates censored)}
}
\value{
The Debrowska bivariate survival function estimate with times for variable 1 along the rows and times for variable 2 along the columns
}
\description{
This takes survival times and censoring indicators for two variables and outputs the Dabrowska bivariate survival function estimate. Modified from code originally written by Yu Cheng at University of Pittsburgh.
}
