% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CauseHaz.R
\name{CauseHaz}
\alias{CauseHaz}
\title{Output the cause specific cumulative hazard and jumps in the cause specific cumulative hazard}
\usage{
CauseHaz(data, cause)
}
\arguments{
\item{data}{An n by 2 matrix where the first column is the observed event time and the second column is the censoring/cause indicator (0 indicates censored)}

\item{cause}{An indicator of which cause to calculate the cause specific hazard for, should be a non-zero value that appears in the cencoring/cause indicator column}
}
\value{
A list with the following elements
\itemize{
\item{hazard: Jumps in the cause specific cumulative hazard estimate}
\item{cumhazard: The cause specific cumulative hazard estimate}
\item{cause: The cause value}
}
}
\description{
This takes a survival time and censoring/cause indicator and outputs the cause specific cumulative hazard and jumps in the cause specific cumulative hazard
}
