% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovComp.R
\name{CovComp}
\alias{CovComp}
\title{Output the martingale and counting process covariance for two variables in presence of competing risks}
\usage{
CovComp(data, t, cause1, cause2)
}
\arguments{
\item{data}{An n by 4 matrix where the first column is the observed event time of the first variable, the second column is the observed event time of the second variable, the third column is the censoring/cause indicator for the first variable (0 indicates censored) and the fourth column is the censoring/cause indicator for the second variable (0 indicates censored)}

\item{t}{The time to evaluate the martingale and counting process covariance at}

\item{cause1}{An indicator of which cause to calculate the cause specific hazard for for the first variable, should be a non-zero value that appears in the censoring/cause indicator column for the first variable}

\item{cause2}{An indicator of which cause to calculate the cause specific hazard for for the second variable, should be a non-zero value that appears in the censoring/cause indicator column for the second variable}
}
\value{
A list with the following elements
\itemize{
\item{MartCov: The martingale covariance between variable 1 and variable 2 at time t}
\item{CountCov: The counting process covariance between variable 1 and variable 2 at timet}
\item{CIFVar1: The cause specifice cumulative incidence function estimate for variable 1}
\item{Survar2: The cause specifice cumulative incidence function estimate for variable 2}
}
}
\description{
This takes a two survival time and  two censoring/cause indicator and calculates the martingale covariance at time t
}
