% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovMatComp.R
\name{CovMatComp}
\alias{CovMatComp}
\title{Output the entire martingale and counting process covariance and correlation matrices for p different variables at time t in the presence of competing risks}
\usage{
CovMatComp(data, causes, p = NA, t, mineigen = 0.001)
}
\arguments{
\item{data}{A matrix where the first p columns are the event times for the p variables and the next p columns are the censoring/cause indicators for the p variables. The order for the event times and censoring/cause indicators needs to be the same.}

\item{causes}{A p-dimensional vector that has the causes for each of the p different variables.}

\item{p}{The number of variables to calculate the matrix for, if NA ncol(data)/2 will be used}

\item{t}{The timepoint to estimate the covariance and correlation matrices}

\item{mineigen}{The minimum eigenvalue for the martingale and counting process covariance matrices if not already positive semidefinite}
}
\value{
A list with the following elements
\itemize{
\item{CovMartComp: The full martingale covariance matrix}
\item{CorMartComp: The full martingale correlation matrix}
\item{CovCountComp: The full counting process covariance matrix}
\item{CorCountComp: The full counting process correlation matrix}
}
}
\description{
This takes survival times and censoring/cause indicators for p different variables and outputs the p times p martingale and counting process covariance/correlation matrices for all variables in the presence of competing risks.
}
