% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovMatNoComp.R
\name{CovMatNoComp}
\alias{CovMatNoComp}
\title{Output the entire martingale and counting process covariance and correlation matrices for p different variables at time t.}
\usage{
CovMatNoComp(data, p = NA, t, mineigen = 0.001)
}
\arguments{
\item{data}{A matrix where the first p columns are the event times for the p variables and the next p columns are the censoring indicators for the p variables. The order for the event times and censoring indicators needs to be the same.}

\item{p}{The number of variables to calculate the matrix for, if NA ncol(data)/2 will be used}

\item{t}{The timepoint to estimate the covariance and correlation matrices}

\item{mineigen}{The minimum eigenvalue for the martingale and counting process covariance matrices if not already positive semidefinite}
}
\value{
A list with the following elements
\itemize{
\item{CovMart: The full martingale covariance matrix}
\item{CorMart: The full martingale correlation matrix}
\item{CovCount: The full counting process covariance matrix}
\item{CorCount: The full counting process correlation matrix}
}
}
\description{
This takes survival times and censoring/causue indicators for p different variables and outputs the p times p martingale and counting process covariance/correlation matrices for all variables.
}
