% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovNoComp.R
\name{CovNoComp}
\alias{CovNoComp}
\title{Output the martingale and counting process covariance for two variables without presence of competing risks}
\usage{
CovNoComp(data, t)
}
\arguments{
\item{data}{An n by 4 matrix where the first column is the observed event time of the first variable, the second column is the observed event time of the second variable, the third column is the censoring indicator for the first variable (0 indicates censored) and the fourth column is the censoring indicator for the second variable (0 indicates censored)}

\item{t}{The time to evaluate the martingale and counting process covariance at}
}
\value{
A list with the following elements
\itemize{
\item{MartCov: The martingale covariance between variable 1 and variable 2 at time t}
\item{CountCov: The counting process covariance between variable 1 and variable 2 at timet}
\item{SurVar1: The Kaplan-Meier survival function estimate for variable 1}
\item{Survar2: The Kaplan-Meier survival funciton estimate for variable 2}
}
}
\description{
This takes survival times and censoring indicators for two variables and outputs the martingale and counting process covariance at time t
}
