% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NelsonAalen.R
\name{NelsonAalen}
\alias{NelsonAalen}
\title{Output the Nelson Aalen cumulative hazard and jumps in the Nelson Aalen cumulative hazard}
\usage{
NelsonAalen(data)
}
\arguments{
\item{data}{An n by 2 matrix where the first column is the observed event time and the second column is the censoring indicator (0 indicates censored)}
}
\value{
A list with the following elements
\itemize{
\item{hazard: Jumps in the Nelson-Aalen estimate for the cumulative hazard}
\item{cumhazard: The Nelson-Aalen estimate for the cumulative hazard}
}
}
\description{
This takes a survival time and censoring indicator and outputs the cumulative hazard and jumps in the cumulative hazard using the Nelson Aalen Estimator
}
